requirements.txt
requests>=2.32.0
python-dotenv==1.0.0
pandas==2.2.0
google-genai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
fastapi>=0.109.0
uvicorn>=0.27.0
